# Apply fused sorted token ids padding