--- a/python/sglang/srt/layers/moe/fused_moe_triton/fused_moe.py
+++ b/python/sglang/srt/layers/moe/fused_moe_triton/fused_moe.py
@@ -752,14 +752,13 @@ def moe_align_block_size(
     sorted_ids = torch.empty(
         (max_num_tokens_padded,), dtype=torch.int32, device=topk_ids.device
     )
-    sorted_ids.fill_(topk_ids.numel())
-
     max_num_m_blocks = triton.cdiv(max_num_tokens_padded, block_size)
     expert_ids = torch.empty(
         (max_num_m_blocks,), dtype=torch.int32, device=topk_ids.device
     )
     num_tokens_post_pad = torch.empty((1), dtype=torch.int32, device=topk_ids.device)
     if enable_moe_align_block_size_triton:
+        sorted_ids.fill_(topk_ids.numel())
         moe_align_block_size_triton(
             topk_ids,
             num_experts,
@@ -777,6 +776,11 @@ def moe_align_block_size(
             dtype=torch.int32,
             device=topk_ids.device,
         )
+        pad_sorted_token_ids = False
+        if sorted_ids.shape[0] <= 4096:
+            pad_sorted_token_ids = True
+        else:
+            sorted_ids.fill_(topk_ids.numel())
 
         sgl_moe_align_block_size(
             topk_ids,
@@ -787,6 +791,7 @@ def moe_align_block_size(
             num_tokens_post_pad,
             token_cnts_buffer,
             cumsum_buffer,
+            pad_sorted_token_ids,
         )
     return sorted_ids, expert_ids, num_tokens_post_pad